ai-ml/use-cases/03_identification_beneficiary/requirements.txt
# Auto Identification of Beneficiaries Requirements
# Use Case ID: AI-PLATFORM-03

# Core Data Science
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
shap>=0.42.0  # For explainability

# MLflow
mlflow>=2.8.0
# Note: mlflow.xgboost is included in mlflow package, no separate installation needed

# Explainability
shap>=0.42.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Configuration
pyyaml>=6.0

# Rule Engine
pyparsing>=3.0.0  # For parsing rule expressions

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3

# Testing (optional)
pytest>=7.4.0
pytest-cov>=4.1.0

# Web Interface (for viewing rules)
flask>=3.0.0

